pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claw-memory"
version = "2.7.1"
description = "AI Memory System with RRF search, knowledge graph, and temporal tracking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Claw Team", email = "claw@example.com"}
]
keywords = ["memory", "ai", "vector-search", "knowledge-graph", "rrf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "lancedb>=0.27.0",
    "pyarrow>=14.0.0",
    "numpy>=1.24.0",
    "networkx>=3.0",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
embedding-ollama = [
    "ollama>=0.1.0",
]
embedding-openai = [
    "openai>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/openclaw/claw-memory"
Documentation = "https://github.com/openclaw/claw-memory#readme"
Repository = "https://github.com/openclaw/claw-memory"

[tool.setuptools.packages.find]
where = ["."]
include = ["memory_*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"
